[/ajax/createAccount.php] - Adapted to work in the "My Company" Tab
[/ajax/deleteAccount.php]
- Adapted to work in the "My Company" Tab

[/ajax/getCompanyData.php]
- fixed possible XSS Attack

[/ajax/getContentForUser.php]
- added Checks for "My Company", "Shell" and "Sensor" Tabs

[/ajax/getUserData.php]
- fixed possible XSS Attack

[/ajax/listCompanies.php]
- fixed possible XSS Attack

[/ajax/Sensors.php]
- added File
- lists all Sensors in the DB and returns JSON Object

[/ajax/listUsers.php]
- Adapted to work in the "My Company" Tab
- fixed possible XSS Attack

[/ajax/updateAccount.php]
- a user cant edit another user with a higher role anymore

[/includes/accounts.html]
- fixed a bug where the user list would not be updated after a Delete action was complete
- has new variable "userRole" which stores the users RoleLvl, to remove delete and edit buttons on Users with higher roles
- can handle more errors visually now

[/includes/accounts.html]
- added handles to work correctly with XSS protection

[/includes/login.html]
- changed Path of CSS file

[/includes/main.html]
- added "Sensors" Tab

[/includes/my account.html]
- added handles to work correctly with XSS protection

[/includes/sensors.html]
- added file
- lists all Sensors in DB

[/includes/shell.php]
- added file
- replaces "System" Tab
- web shell

[/includes/system.php]
- removed file

[/resources/css/includes/companies.css]
- fixed misplacing of #createBox and #editBox

[/resources/css/includes/login.css]
- moved file from /resources/css

[/resources/css/includes/main.css]
- moved file from /resources/css

[/resources/css/includes/sensors.css]
- added file
- contains all styles for "Sensors" Tab

[/scripts/parseURLForQueries.php]
- added file
- saves all queries for the URL, a page has been accessed with, in $queries

[/scripts/replaceSpecialChars.php]
-added file
- replaces "<", ">", " ", "," characters with the HTML charcode

+++ ajax/deleteAccount.php
@@ -1,4 +1,5 @@
 <?php
+
     $silent = true; // if $silent is set to true, getUserData.php wont echo
 
     include_once "../scripts/connectToDatabase.php";
@@ -13,19 +14,37 @@
             $sqlAddition = "";
         }
 
-        if ($conn != false) {
-            // Deleting Dataset
-            $sql = "DELETE FROM tblbenutzer WHERE idIdentifikationsNummer = ?" . $sqlAddition;
-            $stmt = $conn->prepare($sql);
-            $stmt->bind_param(
-                "s",
-                $_POST["id"]
-            );
-            $stmt->execute();
-            $response["SQLError"] = $stmt->error;
+        $sql = "SELECT dtRolle FROM tblBenutzer WHERE idIdentifikationsNummer = ?";
+        $stmt = $conn->prepare($sql);
+        $stmt->bind_param(
+            "i",
+            $_POST["id"]
+        );
+        $stmt->execute();
+        $response["SQLError"] = $stmt->error;
+        $result = $stmt->get_result();
 
-            $response["deletedAccount"] = true;
+        $role = $result->fetch_assoc()["dtRolle"];
+        echo $role;
+
+        if ($response["userData"]["role"] < $role) {
+            $response["deletedAccount"] = false;
+        } else {
+            if ($conn != false) {
+                // Deleting Dataset
+                $sql = "DELETE FROM tblbenutzer WHERE idIdentifikationsNummer = ?" . $sqlAddition;
+                $stmt = $conn->prepare($sql);
+                $stmt->bind_param(
+                    "s",
+                    $_POST["id"]
+                );
+                $stmt->execute();
+                $response["SQLError"] = $stmt->error;
+
+                $response["deletedAccount"] = true;
+            }
         }
+
     }
 
     echo json_encode($response);
